authenticator/requirements/base.txt
fastapi
pydantic-settings
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
uvicorn[standard]
# # linting
# ruff

